Hack to get Mario to train with stable-baselines3
The way the vec_env wrapper code is written, if any custom rewards are
used in the code, they will not be displayed correctly in the monitor
portion as the custom reward wrapper comes after the monitor. To fix
this, the monitor should be called after the custom wrapper is used.

Signed-Off-By: Robert Clark <robdclark@outlook.com>